project/requirements.txt
flask
flask-socketio
openai
python-dotenv
